requirements.txt
smolagents>=0.0.6
click>=8.1.7
python-dotenv>=1.0.0
gitpython>=3.1.42
pyyaml>=6.0.1 requests>=2.31.0
